Completed Tasks (b), (c)
Task (b) Ask me to output to variable "area" to 3 decimal places.

I achieved this by using "%.Xf"
where X is equal to number decimal to places to output.

Task (c) Asks me to output the values of 2 variable "area" and "radius" in a single printf statement

I achieved this by separating each variable with commas and calling each of the variables respectively.